Pull out module init_handle
