+++ packages/std/src/types.rs
@@ -96,66 +96,6 @@ pub struct Coin {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "lowercase")]
-pub enum CosmosMsg {
-    // this moves tokens in the underlying sdk
-    Send {
-        from_address: HumanAddr,
-        to_address: HumanAddr,
-        amount: Vec<Coin>,
-    },
-    // this dispatches a call to another contract at a known address (with known ABI)
-    // msg is the json-encoded HandleMsg struct
-    Contract {
-        contract_addr: HumanAddr,
-        msg: Binary, // we pass this in as Vec<u8> to the contract, so allow any binary encoding (later, limit to rawjson?)
-        send: Option<Vec<Coin>>,
-    },
-    // this should never be created here, just passed in from the user and later dispatched
-    Opaque {
-        data: Binary,
-    },
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "lowercase")]
-pub enum ContractResult {
-    Ok(Response),
-    Err(String),
-}
-
-impl ContractResult {
-    // unwrap will panic on err, or give us the real data useful for tests
-    pub fn unwrap(self) -> Response {
-        match self {
-            ContractResult::Err(msg) => panic!("Unexpected error: {}", msg),
-            ContractResult::Ok(res) => res,
-        }
-    }
-
-    pub fn is_err(&self) -> bool {
-        match self {
-            ContractResult::Err(_) => true,
-            _ => false,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
-pub struct LogAttribute {
-    pub key: String,
-    pub value: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Default, Debug, PartialEq, JsonSchema)]
-pub struct Response {
-    // let's make the positive case a struct, it contrains Msg: {...}, but also Data, Log, maybe later Events, etc.
-    pub messages: Vec<CosmosMsg>,
-    pub log: Vec<LogAttribute>, // abci defines this as string
-    pub data: Option<Binary>,   // abci defines this as bytes
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "lowercase")]
 pub enum QueryResult {
     Ok(Binary),
     Err(String),
@@ -184,47 +124,4 @@ pub fn coin(amount: &str, denom: &str) -> Vec<Coin> {
         amount: amount.to_string(),
         denom: denom.to_string(),
     }]
-}
-
-// log is shorthand to produce log messages
-pub fn log(key: &str, value: &str) -> LogAttribute {
-    LogAttribute {
-        key: key.to_string(),
-        value: value.to_string(),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::serde::{from_slice, to_vec};
-
-    #[test]
-    fn can_deser_error_result() {
-        let fail = ContractResult::Err("foobar".to_string());
-        let bin = to_vec(&fail).expect("encode contract result");
-        println!("error: {}", std::str::from_utf8(&bin).unwrap());
-        let back: ContractResult = from_slice(&bin).expect("decode contract result");
-        assert_eq!(fail, back);
-    }
-
-    #[test]
-    fn can_deser_ok_result() {
-        let send = ContractResult::Ok(Response {
-            messages: vec![CosmosMsg::Send {
-                from_address: HumanAddr("me".to_string()),
-                to_address: HumanAddr("you".to_string()),
-                amount: coin("1015", "earth"),
-            }],
-            log: vec![LogAttribute {
-                key: "action".to_string(),
-                value: "release".to_string(),
-            }],
-            data: None,
-        });
-        let bin = to_vec(&send).expect("encode contract result");
-        println!("ok: {}", std::str::from_utf8(&bin).unwrap());
-        let back: ContractResult = from_slice(&bin).expect("decode contract result");
-        assert_eq!(send, back);
-    }
 }
